// src/interfaces/Dropdown.interface.ts
export interface DropdownInterface {
  inputClass: string;
  Pleacheholder?: string;
  id: string;
  options?: any[];
  onChange?: any;
  name?: string;
  showOptions?: boolean;
  toggleOptions?: any;
  setShowOptions?: any;
  selectedOption?: any;
  setSelectedOption?: any;
  setShowClearButton?: any;
}